fix dtype compare

+++ src/Native/src/kernels/stackvm/optimized/riscv64/binary.cpp
@@ -350,7 +350,7 @@ static int verify_shape_impl(gsl::span<const size_t> in_a_shape,
 
     if (in_b_shape[in_b_shape.size() - 1] == 1) {
         int len_b_leave = 1;
-        for (int i = index + 1; i < in_b_shape.size(); ++i) {
+        for (int i = index + 1; i < (int)in_b_shape.size(); ++i) {
             len_b_leave *= in_b_shape[i];
         }
         if (len_b_leave != 1) {
@@ -392,7 +392,7 @@ static int verify_shape(gsl::span<const size_t> in_a_shape,
 static gsl::span<const size_t>
 get_sample_span(gsl::span<const size_t> in_shape) {
     int not_one_index = 0;
-    for (int i = 0; i < in_shape.size(); ++i) {
+    for (int i = 0; i < (int)in_shape.size(); ++i) {
         if (in_shape[i] != 1) {
             not_one_index = i;
             break;
@@ -439,7 +439,7 @@ get_sample_span(gsl::span<const size_t> in_shape) {
                     (index != (int)(in_b_shape.size() - 1))) {                 \
                     int size_diff = in_a_shape.size() - in_b_shape.size();     \
                     int len_a_leave = 1;                                       \
-                    for (int i = index + 1; i < in_b_shape.size(); ++i) {      \
+                    for (int i = index + 1; i < (int)in_b_shape.size(); ++i) {      \
                         len_a_leave *= in_a_shape[i + size_diff];              \
                     }                                                          \
                     for (int j = 0; j < outter_front_size; ++j) {              \
@@ -470,7 +470,7 @@ get_sample_span(gsl::span<const size_t> in_shape) {
                     (index != (int)(in_a_shape.size() - 1))) {                 \
                     int size_diff = in_b_shape.size() - in_a_shape.size();     \
                     int len_b_leave = 1;                                       \
-                    for (int i = index + 1; i < in_a_shape.size(); ++i) {      \
+                    for (int i = index + 1; i < (int)in_a_shape.size(); ++i) {      \
                         len_b_leave *= in_b_shape[i + size_diff];              \
                     }                                                          \
                     for (int j = 0; j < outter_front_size; ++j) {              \
